Controlado el error

+++ api/signals.py
@@ -8,6 +8,7 @@ from django.dispatch import receiver
 from common.models.achievement import UserAchievementProgress, Achievement
 from common.models.route import Route
 from common.models.user import User
+from common.models.calendar import GoogleOAuth2Token
 from .service.calendar import add_event_calendar, delete_event_calendar
 import datetime
 
@@ -79,7 +80,7 @@ def check_and_increment_progress(user_achievement, achievement, instance):
 @receiver(post_save, sender=Route)
 def route_created_driver_calendar_event(sender, instance, created, **kwargs):
     if created:
-
+        
         if isinstance(instance.duration, int):
             instance.duration = datetime.timedelta(seconds=instance.duration)
 
@@ -100,27 +101,34 @@ def route_created_driver_calendar_event(sender, instance, created, **kwargs):
             },
         }
 
-        # Try to add the event to the driver's calendar
-        try:
-            add_event_calendar(instance.driver, instance, event_data)
-        except Exception:
-            pass
+        token_exist = GoogleOAuth2Token.objects.filter(user=instance.driver).exists()
+
+        if token_exist:
+            try:
+                add_event_calendar(instance.driver, instance, event_data)
+            except Exception:
+                pass
 
 
 # Delete the calendar event to the driver when a route is cancelled
 @receiver(post_save, sender=Route)
 def route_cancelled_driver_calendar_event(sender, instance, created, **kwargs):
     if not created and instance.cancelled:
-        try:
-            delete_event_calendar(instance.driver, instance)
-        except Exception:
-            pass
+        token_exist = GoogleOAuth2Token.objects.filter(user=instance.driver).exists()
 
-        for passenger in instance.passengers.all():
+        if token_exist:
             try:
-                delete_event_calendar(passenger, instance)
+                delete_event_calendar(instance.driver, instance)
             except Exception:
                 pass
+
+            for passenger in instance.passengers.all():
+                token_exist = GoogleOAuth2Token.objects.filter(user=passenger).exists()
+                if token_exist:
+                    try:
+                        delete_event_calendar(passenger, instance)
+                    except Exception:
+                        pass
 
 
 # Create a calendar event to the passengers when they join a route
@@ -150,11 +158,13 @@ def passenger_join_route_calendar_event(sender, instance, action, pk_set, **kwar
         }
 
         for user_id in pk_set:
-            try:
-                user = User.objects.get(pk=user_id)
-                add_event_calendar(user, instance, event_data)
-            except User.DoesNotExist:
-                pass
+            token_exist = GoogleOAuth2Token.objects.filter(user=user_id).exists()
+            if token_exist:
+                try:
+                    user = User.objects.get(pk=user_id)
+                    add_event_calendar(user, instance, event_data)
+                except User.DoesNotExist:
+                    pass
 
 
 # Delete the calendar event to the passengers when they leave a route
@@ -164,8 +174,10 @@ def passenger_leave_route_calendar_event(sender, instance, action, pk_set, **kwa
     if action == "post_remove":
         
         for user_id in pk_set:
-            try:
-                user = User.objects.get(pk=user_id)
-                delete_event_calendar(user, instance)
-            except User.DoesNotExist:
-                pass
+            token_exist = GoogleOAuth2Token.objects.filter(user=user_id).exists()
+            if token_exist:
+                try:
+                    user = User.objects.get(pk=user_id)
+                    delete_event_calendar(user, instance)
+                except User.DoesNotExist:
+                    pass
